requirements.txt
google-cloud-bigtable
pandas
prophet
plotly
